fix(types): Move survey types back to main file
